requirements.txt
mysqlclient==2.1.1
django-environ==0.9.0
pika==1.3.1
celery==5.2.7
django-mailgun-mime==0.1.7
redis==4.3.4
